Fix error in clunit_name definition
The diphone names weren't being used due to an error in lvl_is_v0_clunits.scm. This caused unit distance tables to become extremely large, creating a bottleneck (which seemed infinite). This pull request should fix that error. We might want to experiment with a couple of different setups for these clunit names.